backend/requirements.txt
flask
transformers
torch
sentence-transformers
faiss-cpu
pypdf2
dotenv